test(integration): move migrations-path helpers into tagged bridges
migrationsPathForUnit and migrationsPathForTest are now referenced only
by the tagged TestMain bridges, so under the no-tag lint/build they were
flagged unused. Move each into its bridge file (compiled only under the
tag). Also fix templateHash to satisfy both errcheck and staticcheck
(discard fmt.Fprintf returns to a hash.Hash, which never errors).
